moved task model into their own package
