--- src/main/java/com/geraldsaccount/neuefische_todo/model/dto/TaskDTO.java
@@ -1,7 +1,0 @@
-package com.geraldsaccount.neuefische_todo.model.dto;
-
-import com.geraldsaccount.neuefische_todo.model.TaskStatus;
-
-public record TaskDTO(String description, TaskStatus status) {
-
-}
