+++ src/main/java/com/geraldsaccount/neuefische_todo/model/tasks/dto/TaskDTO.java
@@ -1,0 +1,7 @@
+package com.geraldsaccount.neuefische_todo.model.tasks.dto;
+
+import com.geraldsaccount.neuefische_todo.model.tasks.TaskStatus;
+
+public record TaskDTO(String description, TaskStatus status) {
+
+}
