+++ src/main/java/com/geraldsaccount/neuefische_todo/repository/TaskRepo.java
@@ -3,7 +3,7 @@ package com.geraldsaccount.neuefische_todo.repository;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import com.geraldsaccount.neuefische_todo.model.Task;
+import com.geraldsaccount.neuefische_todo.model.tasks.Task;
 
 @Repository
 public interface TaskRepo extends MongoRepository<Task, String> {
