+++ src/main/java/com/geraldsaccount/neuefische_todo/service/GlobalExceptionHandler.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import com.geraldsaccount.neuefische_todo.model.dto.ErrorDTO;
+import com.geraldsaccount.neuefische_todo.model.error.ErrorDTO;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
